+++ lib/pragma/operation.rb
@@ -16,6 +16,7 @@ require 'pragma/operation/response/unprocessable_entity'
 require 'pragma/operation/response/created'
 require 'pragma/operation/response/ok'
 require 'pragma/operation/response/no_content'
+require 'pragma/operation/response/unauthorized'
 
 module Pragma
   # Operations provide business logic encapsulation for your JSON API.